+++ db/dbinit.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "Users" (
-  "id" SERIAL PRIMARY KEY,
+  "id" INT NOT NULL PRIMARY KEY,
   "login" VARCHAR(255),
   "nickname" VARCHAR(255),
   "avatar" VARCHAR(255)
@@ -55,7 +55,6 @@ CREATE TABLE "MessagesReactions" (
 CREATE TABLE "Themes" (
   "id" SERIAL PRIMARY KEY,
   "name" VARCHAR(255),
-  "description" TEXT NOT NULL,
   UNIQUE ("name")
 );
 
@@ -79,3 +78,5 @@ CREATE INDEX idx_messagesreactions_message ON "MessagesReactions" ("message_id")
 CREATE INDEX idx_messagesreactions_reaction ON "MessagesReactions" ("reaction_id");
 CREATE INDEX idx_themes_id ON "Themes" ("id");
 CREATE INDEX idx_usersthemes_user ON "UsersThemes" ("user_id");
+
+INSERT INTO "Themes" (id, name) VALUES (1, 'light'), (2, 'dark');
