+++ packages/server/api/models/index.ts
@@ -1,0 +1,5 @@
+import {User} from './user';
+
+export {
+	User,
+};
